% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_email_addresses.R
\name{get_email_addresses}
\alias{get_email_addresses}
\title{Get email address from github username}
\usage{
get_email_addresses(github_username)
}
\arguments{
\item{github_username}{A string containing someone's github username.}
}
\value{
A character vector (probably of length 1) containing the email
address of that user.
}
\description{
Gets the email address of a person based upon their github username.
}
\details{
This currently trawls the whole of the user's github public event
data for strings that look like email addresses. that means that other
email addresses beyond the one belonging to the github user could be found
(though this seems to be a rare thing).
}
\examples{
get_email_addresses("richierocks")
}
